+++ src/main/java/com/transmetro/services/BusesSvc.java
@@ -17,11 +17,11 @@ public interface BusesSvc {
 
     List<Buses> obtenerTodos();
 
-    Optional<Buses> obtenerPorId(Integer id);
+    Optional<Buses> obtenerPorId(Long id);
 
     Buses crear(Buses bus);
 
-    Optional<Buses> actualizar(Integer id, Buses datos);
+    Optional<Buses> actualizar(Long id, Buses datos);
 
-    boolean eliminar(Integer id);
+    boolean eliminar(Long id);
 }
